[Refactor] Simplify error dialog formatting
